Refactor serializers
